requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astroid @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b7z3htqpm1/croot/astroid_1717618370968/work
better-rich-prompts==1.0.2
black==24.4.2
Brotli @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_27zk0eqdh0/croot/brotli-split_1714483157007/work
build==1.2.1
CacheControl==0.14.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
dill @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_28zwy_olqk/croot/dill_1715094676263/work
distlib==0.3.8
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.1.1
fastjsonschema==2.19.1
filelock==3.14.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
installer==0.7.0
isort @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_1bnvhw7_ex/croot/isort_1718291367347/work
jaraco.classes==3.4.0
keyring==24.3.1
markdown-it-py==3.0.0
mccabe @ file:///opt/conda/conda-bld/mccabe_1644221741721/work
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.8
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.10.0
platformdirs @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/platformdirs_1701803010714/work
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pylint @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b623auwyzd/croot/pylint_1717622581741/work
pyproject_hooks==1.1.0
PySocks @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pysocks_1699237568675/work
python-dateutil==2.9.0.post0
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv_1662230030282/work
pytz @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a4b76c83ik/croot/pytz_1713974318928/work
rapidfuzz==3.9.1
reclaim-sdk @ git+https://github.com/cato447/reclaim-sdk.git@dd1bd0792be24b06fe57f40633a197699f7888c0
requests @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_b3tnputioh/croot/requests_1707355573919/work
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
things.py @ git+https://github.com/cato447/things.py.git@10fa2c8a07142cd149ef59bff22931ae0856c114
toggl_python @ file:///Users/cato/Code/Cato447/toggl-python
toml==0.10.2
tomlkit @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/tomlkit_1699238737474/work
trove-classifiers==2024.5.22
typer==0.12.3
typing_extensions==4.12.2
urllib3 @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_22oz53beet/croot/urllib3_1715635830593/work
virtualenv==20.26.2
wheel==0.43.0
xattr==1.1.0
